requirements.txt
pygame>=2.1.0
Flask>=2.0.0
opencv-python-headless>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0